% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_proteins.R
\name{append_proteins}
\alias{append_proteins}
\title{Appends proteins to a cluster they derrive from}
\usage{
append_proteins(rebased_clusters)
}
\arguments{
\item{rebased_clusters}{input cluster file as a list but trimmed to the current
iteration.}
}
\value{
a character vector of protein descriptions
}
\description{
Appends proteins to a cluster they derrive from
}
