% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_faa_files.R
\name{detect_faa_files}
\alias{detect_faa_files}
\title{detect_faa_files
Detects .faa files in the specified directory}
\usage{
detect_faa_files(faa_fdir)
}
\arguments{
\item{faa_fdir}{directory with .faa files}
}
\value{
character vector of .faa files
}
\description{
detect_faa_files
Detects .faa files in the specified directory
}
