% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ref_prot.R
\name{find_ref_prot}
\alias{find_ref_prot}
\title{find_ref_prot
Finds line with the reference protein within a cluster}
\usage{
find_ref_prot(single_cluster = x)
}
\arguments{
\item{single_cluster}{a single cluster as a named list}
}
\value{
character vector with the reference protein line
}
\description{
find_ref_prot
Finds line with the reference protein within a cluster
}
