% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_clusters.R
\name{gather_clusters}
\alias{gather_clusters}
\title{Gather clusters}
\usage{
gather_clusters(clstr_file)
}
\arguments{
\item{clstr_file}{output from CD-HIT program, as a list (each line as an element)}
}
\description{
Function processes flat .clstr file - an output from CD-HIT program, and
creates a list where each element is a protein cluster name and sub elements
are descriptions of proteins and alignments
}
