% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_prot_details.R
\name{get_ref_prot_details}
\alias{get_ref_prot_details}
\title{Extracts organism/genome name as well as protein identifier from reference protein line}
\usage{
get_ref_prot_details(ref_prot_line, genomes)
}
\arguments{
\item{ref_prot_line}{character vector; must contain genome name (equal to the
.faa file name), protein identifier must start with ">" and be separated
from genome name with underscore "_"}

\item{genomes}{character vector with genome names taken from .faa files}
}
\value{
list with genome name and protein identifier of the reference protein
}
\description{
Extracts organism/genome name as well as protein identifier from reference protein line
}
\examples{
get_ref_prot_details(">PROTEIN_ID1_GENOME_A", c("GENOME_A", "GENOME_B"))

}
