\name{papiHtest}
\alias{papiHtest}

\title{
Apply ANOVA or t-test to results produced by \code{\link{papi}}.
}
\description{
papiHtest applies ANOVA or t-test to a data frame generated by papi.

}
\usage{
papiHtest(inputData, signif.level = 0.05, log.transform = TRUE, 
save = TRUE, folder, StatTest, output, adjust.pValue = TRUE, 
method = "bonferroni")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{inputData}{
when inputData is missing, a dialog box will pop up allowing the user to click-and-point to the comma separated value (CSV) file from which the data is to be read. Alternatively, inputData can take a character string naming the path to the CSV file to be read or the name of a variable (data frame) containing the data frame resultant from \code{\link{papi}}.
}
  \item{signif.level}{
A numeric value indicating the p-value cut-off. Every metabolic pathway showing a p-value higher than the signif.level will be removed from the data frame produced by papiHtest.  
}
  \item{log.transform}{
A logical vector (TRUE or FALSE) defining if the data should be log-transformed before applying the statistical test.
}
  \item{save}{
A logical vector (TRUE or FALSE) defining if the results must be saved into a CSV file.
}
  \item{folder}{
When save = TRUE and folder is missing, a pop up dialog box will be presented to the user. The user can then select the directory to which the results will be saved. Alternatively, folder can take a character string naming the path to the folder where the results must be saved.
}
  \item{StatTest}{
When StatTest is missing, a t-test is performed if the input data has exactly two experimental conditions and an ANOVA is performed otherwise. Alternatively, StatTest may receive the values "ANOVA", "Anova", "anova", "A" or "a" for ANOVA, or "T-TEST", "T-test", "t-test", "t-TEST", "t" and "T" for t-test.  
}
  \item{output}{
A character string indicating the name of the file containing the results. When output is missing, the CSV file produced is named according to the statistical test performed, t-test.csv or ANOVA.csv. 
}
  \item{adjust.pValue}{
A logical vector indicating if p-values should be adjusted by the \code{\link{p.adjust}}.
}
  \item{method}{
A character string indicating the method used when applying \code{\link{p.adjust}}. See \code{\link{p.adjust.methods}} for possible methods. Default is 'bonferroni'.
}
}
\details{
The inputData must have the same format as data(papiResults). papiHtest can be applied leaving all of it arguments as default. In this case, pop up windows will let the user choose the required input data file and the folder to save results.     
}
\value{
papiHtest generates a data frame as the input data, however, containing an additional column with the calculated p-values. Metabolic pathways showing a p-value higher than the signif.level are removed.
}
\references{
Aggio, R.B.M; Ruggiero, K. and Villas-Boas, S.G. (2010) - Pathway Activity Profiling (PAPi): from metabolite profile to metabolic pathway activity. Bioinformatics.
}
\author{
Raphael Aggio
}
\note{
Raphael Aggio (raphael.aggio@gmail.com)
}

\seealso{
\code{\link{buildDatabase}}, \code{\link{papi}}, \code{\link{papiLine}} and \code{\link{addKeggCodes}}.
}
\examples{
### Building input data ####
Names <- c("Replicates", "Galactose metabolism", "Glycerolipid metabolism", 
"Carbon fixation in photosynthetic organisms", "Fatty acid biosynthesis", 
"D-Alanine metabolism")
Sample1 <- c("cond1", 200, 77, 99, 42, 29)
Sample2 <- c("cond1", 236, 72, 80, 39, 14)
Sample3 <- c("cond1", 269, 83, 89, 45, 31)
Sample4 <- c("cond2", 68, 12, NA, 6, 0.006)
Sample5 <- c("cond2", 57, 10, NA, 7, 0.1)
Sample6 <- c("cond2", 58, 13, NA, 8, 0.05) 

dataInput <- data.frame(cbind(Names, Sample1, Sample2, Sample3, Sample4, 
Sample5, Sample6), stringsAsFactors = FALSE)

### Applying papiHtest ####
papiHtestResults <- papiHtest(dataInput, save = FALSE)
}