% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ref_prot.R
\name{remove_ref_prot}
\alias{remove_ref_prot}
\title{Removes reference proteins from cluster}
\usage{
remove_ref_prot(clstr_prot_list, reference_proteins)
}
\arguments{
\item{clstr_prot_list}{list of protein clusters}

\item{reference_proteins}{list with reference proteins in cluster}
}
\value{
list of proteins that are not reference proteins
}
\description{
Removes reference proteins from cluster
}
